requirements.txt
Django==2.1
django-rest-framework==0.1.0
djangorestframework==3.8.2
Pillow==5.2.0
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
reportlab==3.5.5
